pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "openai>=1.68.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "werkzeug>=3.1.3",
]
